LayoutTests/media/modern-media-controls/start-button/start-button-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<style type="text/css" media="screen">

#container {
    position: absolute;
    left: 0;
    top: 0;
    width: 300px;
    height: 200px;
}

</style>
<body>
<div id="container"></div>
<script type="text/javascript">

description("Testing the <code>StartButton</code> class.");

const startButton = new StartButton({ layoutTraits: LayoutTraits.macOS });
document.getElementById("container").appendChild(startButton.element);

const startButtonStyle = window.getComputedStyle(startButton.element);
const startButtonBounds = startButton.element.getBoundingClientRect();
shouldBeEqualToString("startButtonStyle.position", "absolute");
shouldBeEqualToString("startButtonStyle.left", "50%");
shouldBeEqualToString("startButtonStyle.top", "50%");
shouldBeEqualToString("startButtonStyle.width", "70px");
shouldBeEqualToString("startButtonStyle.height", "70px");
shouldBeEqualToString("startButtonStyle.transform", "matrix(1, 0, 0, 1, -35, -35)");
shouldBe("startButtonBounds.left", "115");
shouldBe("startButtonBounds.top", "65");

debug("");
const startButtonBackgroundStyle = window.getComputedStyle(startButton.element.firstElementChild);
shouldBeEqualToString("startButtonBackgroundStyle.position", "absolute");
shouldBeEqualToString("startButtonBackgroundStyle.left", "0px");
shouldBeEqualToString("startButtonBackgroundStyle.top", "0px");
shouldBeEqualToString("startButtonBackgroundStyle.width", "70px");
shouldBeEqualToString("startButtonBackgroundStyle.height", "70px");
shouldBeEqualToString("startButtonBackgroundStyle.webkitAppearance", "media-controls-light-bar-background");
shouldBeEqualToString("startButtonBackgroundStyle.webkitClipPath", "circle(50% at 50% 50%)");
shouldBeEqualToString("startButtonBackgroundStyle.willChange", "transform");

debug("");
const startButtonImageStyle = window.getComputedStyle(startButton.element.lastElementChild);
shouldBeEqualToString("startButtonImageStyle.position", "absolute");
shouldBeEqualToString("startButtonImageStyle.left", "0px");
shouldBeEqualToString("startButtonImageStyle.top", "0px");
shouldBeEqualToString("startButtonImageStyle.width", "70px");
shouldBeEqualToString("startButtonImageStyle.height", "70px");
shouldBeCloseTo(startButtonImageStyle.opacity, 0.6, 0.001);

startButton.element.remove();
debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
